Test and move storeToken.ts

+++ server/utils/database/storeToken/storeToken.ts
@@ -1,6 +1,6 @@
 import type { SupabaseClient } from '@supabase/supabase-js'
 import type { Database } from '~/types/database.types.ts'
-import { deleteToken } from '~/server/utils/database/deleteToken.ts'
+import { deleteToken } from '~/server/utils/database/deleteToken'
 
 export const storeToken = async (tokenData: Database['public']['Tables']['oauth_tokens']['Insert'], client: SupabaseClient<Database>) => {
   const { data, error } = await client.from('oauth_tokens').insert(tokenData)
@@ -15,6 +15,12 @@ export const storeToken = async (tokenData: Database['public']['Tables']['oauth_
 }
 
 export const safeStoreToken = async (tokenData: Database['public']['Tables']['oauth_tokens']['Insert'], client: SupabaseClient<Database>) => {
+  if (!tokenData.user_id || !tokenData.provider) {
+    throw createError({
+      statusCode: 400,
+      statusMessage: 'Missing required fields, user_id and provider are required',
+    })
+  }
   await deleteToken(tokenData.user_id, tokenData.provider, client)
   return storeToken(tokenData, client)
 }
